+++ svc/GridExportService.js
@@ -9,7 +9,7 @@ import {HoistService, XH} from '@xh/hoist/core';
 import {FieldType} from '@xh/hoist/data';
 import {fmtDate} from '@xh/hoist/format';
 import {Icon} from '@xh/hoist/icon';
-import {SECONDS} from '@xh/hoist/utils/datetime';
+import {isLocalDate, SECONDS} from '@xh/hoist/utils/datetime';
 import {throwIf, withDefault} from '@xh/hoist/utils/js';
 import download from 'downloadjs';
 import {StatusCodes} from 'http-status-codes';
@@ -25,7 +25,7 @@ import {
     sortBy,
     compact,
     findIndex,
-    keys
+    keys, isBoolean, isDate, isInteger, isNumber, last, isObject
 } from 'lodash';
 import {span, a} from '@xh/hoist/cmp/layout';
 import {wait} from '@xh/hoist/promise';
@@ -181,13 +181,15 @@ export class GridExportService extends HoistService {
 
         if (isNil(value)) return null;
 
-        // 1) Support per-cell excelFormat
+        // 1) Support per-cell excelFormat and data types
         let {excelFormat} = column,
             cellHasExcelFormat = isFunction(excelFormat);
 
         if (cellHasExcelFormat) {
             excelFormat = excelFormat(value, {record, column, gridModel}) ?? ExcelFormat.DEFAULT;
         }
+
+        const cellSpecificType = this.getTypeIfDifferent(value, this.getExportFieldType(column));
 
         // 2) Dates: Provide the date data string expected by the server endpoint
         // Also functions as a consistent human-friendly date format for CSV and clipboard
@@ -196,9 +198,12 @@ export class GridExportService extends HoistService {
         value = value.toString();
 
         // Send format and/or type with the cell in an object only if it varies within the column
-        return forExcel && cellHasExcelFormat ?
-            {value, format: excelFormat} :
-            value;
+        if (!cellSpecificType && !cellHasExcelFormat) return value;
+
+        const ret = {value};
+        if (cellHasExcelFormat) ret.format = excelFormat;
+        if (cellSpecificType) ret.type = cellSpecificType;
+        return ret;
     }
     //-----------------------
     // Implementation
@@ -380,6 +385,23 @@ export class GridExportService extends HoistService {
             await minDelay;
             fn();
         };
+    }
+
+    // Return a value's data type if different from the type specified
+    getTypeIfDifferent(v, type) {
+        const ifTypeNot = (...types) => types.includes(type) ? null : last(types),
+            {BOOL, INT, NUMBER, LOCAL_DATE, DATE, STRING, PWD, AUTO, TAGS, JSON} = FieldType;
+
+        if (isBoolean(v))   return ifTypeNot(BOOL);
+        if (isInteger(v))   return ifTypeNot(NUMBER, INT);
+        if (isNumber(v))    return ifTypeNot(NUMBER);
+        if (isLocalDate(v)) return ifTypeNot(LOCAL_DATE);
+        if (isDate(v))      return ifTypeNot(DATE);
+        if (isString(v))    return ifTypeNot(PWD, STRING, AUTO);
+        if (isArray(v))     return ifTypeNot(TAGS, JSON, AUTO);
+        if (isObject(v))    return ifTypeNot(JSON, AUTO);
+
+        return null;
     }
 }